requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scipy>=1.7.0
gymnasium>=0.26.0
torch>=1.10.0
stable-baselines3>=1.6.0
pyserial>=3.5
pyyaml>=6.0
scikit-learn>=1.0.0
